+++ client/src/components/CustomizedDelivery/DeliverySummary.tsx
@@ -1,4 +1,4 @@
-import styles from './Summary.module.css';
+import styles from './DeliverySummary.module.css';
 
 export const Summary = () => {
 	return <div className={styles.container}>Summary</div>;
